\name{defraction}
\alias{defraction}
\title{Converts fractional read counts to integers.}
\usage{
defraction(raw)
}
\description{
Convert a matrix of fractional values to integers (always rounds down).}
\arguments{
\item{raw}{A numeric \code{data.frame}}
}
\value{
Returns a \code{data.frame} of integers.
}
\examples{
# defraction(raw)
}
