\docType{data}
\name{test_counts}
\alias{test_counts}
\title{Testing RNAseq raw counts.}
\format{
A data frame with 25239 observations on the following 12 variables.
\describe{
 \item{\code{ctl1_Young}}{numeric, raw counts}
 \item{\code{ctl2_Young}}{numeric, raw counts}
 \item{\code{ctl3_Young}}{numeric, raw counts}
 \item{\code{ctl1_Old}}{numeric, raw counts}
 \item{\code{ctl2_Old}}{numeric, raw counts}
 \item{\code{ctl3_Old}}{numeric, raw counts}
 \item{\code{trt1_Young}}{numeric, raw counts}
 \item{\code{trt2_Young}}{numeric, raw counts}
 \item{\code{trt3_Young}}{numeric, raw counts}
 \item{\code{trt1_Old}}{numeric, raw counts}
 \item{\code{trt2_Old}}{numeric, raw counts}
 \item{\code{trt3_Old}}{numeric, raw counts}
}
}
\source{
Jittered values from true RNAseq samples.
}
\usage{
test_counts
}
\description{
Semi-simulated RNAseq raw counts for testing functionality.
}
\keyword{datasets}
