\docType{data}
\name{test_design}
\alias{test_design}
\title{Meta data for test_counts.}
\format{
A data frame with 12 observations on the following 3 variables.
\describe{
 \item{\code{treatment}}{factor, "ctl" or "trt"}
 \item{\code{age}}{factor, "young" or "old"}
 \item{\code{group}}{factor, age + treatment}
}
}
\source{
Meta data for generated samples.
}
\usage{
test_design
}
\description{
Meta data for test_counts for testing functionality.
}
\keyword{datasets}
